fixed the appointment-form styling in contact-section
Improved the layout of the gender input and swine age input fields for better responsiveness from phone to desktop view. Adjusted the footer padding to maintain proper spacing between the form and the footer on phone and tablet view. Also fixed the overflowing height of the form on desktop view.